requirements.txt
torch
torchvision
pillow
sounddevice
numpy
matplotlib
librosa
